requirements.txt
Flask==0.10.1
py-ninja==0.1.0
